requirements.txt
asgiref==3.8.1
Django==4.2.17
djangorestframework==3.16.0
mysqlclient==2.2.7
PyMySQL==1.1.1
sqlparse==0.5.3
tzdata==2025.2
